+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 8.3
 --------------
 - Fix PHP Warning non-numeric value encountered, use rounded percent grade in StudentGrades.php
+- Security Fix reflected XSS: encode HTML special chars for search_term in Courses.php
 
 Changes in 8.2.1
 ----------------