+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 8.3
 --------------
 - Fix PHP Warning non-numeric value encountered, use rounded percent grade in StudentGrades.php
 - Security Fix reflected XSS: encode HTML special chars for search_term in Courses.php
+- Add File Attached to Staff Payments in StaffPayments.php & Accounting/functions.inc.php
 
 Changes in 8.2.1
 ----------------